# proto/openapi/console.yaml
...
openapi: 3.1.0
info:
  title: AIRush Console API
  version: v1
  description: 控制面领域模型 CRUD（datasources/agents/groups/aliases/connectors）
...
paths:
...
  /api/v1/connectors/{id}:
    get:
      summary: Connector 详情
      operationId: getConnector
      tags: [connectors]
      parameters: [{ $ref: "#/components/parameters/ID" }]
      responses:
        "200":
          description: 详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Connector" }
        default: { $ref: "#/components/responses/Error" }
...
components:
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Error:
      description: 标准错误响应（spec-0.8 契约）
      content:
        application/json:
          schema: { $ref: "./error.yaml#/components/schemas/ErrorResponse" }
  schemas:
...
    Connector:
      type: object
      required:
        [id, name, location, version, status, last_heartbeat_at, cert_fingerprint,
         created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        location: { type: string }
        version: { type: string }
        status: { type: string, enum: [online, offline, degraded] }
        last_heartbeat_at: { type: [string, "null"], format: date-time }
        cert_fingerprint: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...